feat: implement mapping for job throw error request
- Refactored RequestMapper class, moved validators into separated classes
